backend/requirements.txt
# FastAPI 核心
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# 数据库
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Redis
redis==5.0.1

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 数据验证
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# 工具
python-dotenv==1.0.1
httpx==0.26.0

# 测试
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0

# AI Agent
openai>=1.0.0             # 智谱 API (OpenAI 兼容)
openai-whisper>=20231117  # 语音识别 (可选，需要 ffmpeg)
